pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytsum"
version = "0.1.0"
description = "YouTube Transcript Summarizer - Automated video summarization with AI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "textual>=0.47.0",
    "youtube-transcript-api>=0.6.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "schedule>=1.2.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
ytsum = "ytsum.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]
